Logic for handling typdef enum declarations
Prior to this commit a specific type of C typedef declaration broke
this tool's functionality.

```
enum tux {
    aa,
    bb,
    cc
};

typedef enum tux tux;
```

Running autopxd on a file contaning this code resulted in an
`'AttributeError: 'NoneType' object has no attribute 'enumerators''`
error and failure to write the pxd file. After these changes the
tool does not generate a ctypedef in this scenario. Initially
only changes to the `visitEnum` function were incorporated which
resulted in a `ctypedef tux tux` declaration but that caused an
incomplete / ambiguous struct error when compiling the associated
pyx file. After these changes this tool will generate the following
in a pxd file associated with the above code snippet.

```
cdef enum tux:
    aa
    bb
    cc
```