Catch macOS clang error
The clang version that ships with the latest XCode requires a quoted path input to it's `-I` option. This patch is a bit basic – it assumes that if you're running under macOS (`darwin`), that you're using the XCode toolchain, and thus clang.
I'm happy to submit something more robust, but I'm not sure how to do it besides catching the error and retrying.