Added comments and logic for whitelist generation
As a user I've had the desire to generate a .pxd file that
contained mappings from only a few select files. Prior to
this commit it was unclear how to accomplish that. After
this commit a user can modify the source code to utilize
a 'whitelist' of only the files they would like to generate
mappings from.

By defualt the whitelist is not populated but comments were
added so that users with interests simialr to mine might be
pointed in the right direction in the future.