Added logic for whitelist generation.
As a user I've had the desire to generate a .pxd file that
contained mappings from only a few select files. Prior to
this commit it was unclear how to accomplish that. After
this commit a user can utilize a 'whitelist' of only the
files they would like to generate mappings from.

This commit exposes the whitelist functionality to programs that
use the autopxd module through an optional parameter to the translate
function. Further a special case test was added to the test suite along
with test headers required for the functionality to be exercised.